function-slideshow-creator/requirements.txt
moviepy
google-cloud-storage
functions-framework